requirements.txt
openai==1.29.0
anthropic==0.25.7
google-generativeai==0.5.2
tiktoken==0.7.0
streamlit==1.33.0
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
langchain-core==0.1.46
langchain-google-genai==1.0.3
langchain-anthropic==0.1.11
python-dotenv==1.0.0

# Webサイト要約アプリ用
requests==2.31.0
beautifulsoup4==4.12.3
langchain_text_splitters==0.0.1

# Youtube要約アプリ用
youtube-transcript-api==0.6.2
pytube==15.0.0


PyMuPDF==1.23.25
faiss-cpu==1.7.4


langsmith==0.1.54


duckduckgo_search==4.2
html2text==2020.1.16
lxml[html_clean]==5.2.1
readability-lxml==0.8.1


streamlit-feedback==0.1.3

python-magic==0.4.27
google-cloud-bigquery==3.21.0
db-dtypes==1.2.0
